+++ src/components/TrackerCardList/TrackerCardList.tsx
@@ -1,7 +1,8 @@
+import { useAutoAnimate } from '@formkit/auto-animate/react';
 import { CardProps, useTheme } from '@mui/material';
 import { FC } from 'react';
 
-import { useAppSelector } from '../../app/hooks';
+import { useAppSelector } from '../../hooks/redux';
 import Tracker from '../../models/Tracker';
 import { selectThemeMode } from '../../store/theme/theme.selectors';
 import TrackerCard from '../TrackerCard/TrackerCard';
@@ -12,20 +13,21 @@ interface Props {
   trackers: Tracker[];
 }
 
-const TrackerList: FC<Props> = ({ trackers, cardProps }) => {
+const TrackerCardList: FC<Props> = ({ trackers, cardProps }) => {
   const themeMode = useAppSelector(selectThemeMode);
   const theme = useTheme();
+  const [animateRef] = useAutoAnimate<HTMLDivElement>();
   const allCardProps = {
     ...defaultCardProps(themeMode, theme),
     ...cardProps
   };
   return (
-    <>
+    <div ref={animateRef}>
       {trackers.map((t) => (
         <TrackerCard tracker={t} key={t.id} cardProps={allCardProps} />
       ))}
-    </>
+    </div>
   );
 };
 
-export default TrackerList;
+export default TrackerCardList;