+++ src/components/forms/NumberTextField/NumberTextField.tsx
@@ -2,15 +2,17 @@ import { TextField, TextFieldProps } from '@mui/material';
 import { FC } from 'react';
 
 const NumberTextField: FC<TextFieldProps> = (textFieldProps) => {
+  const { inputProps, ...otherProps } = textFieldProps;
   return (
     <TextField
       inputProps={{
         inputMode: 'numeric',
         min: 1,
         pattern: '[0-9]*',
-        type: 'number'
+        type: 'number',
+        ...inputProps
       }}
-      {...textFieldProps}
+      {...otherProps}
     />
   );
 };