+++ src/components/forms/ValidateCompletionsForm/ValidateCompletionsForm.tsx
@@ -1,8 +1,9 @@
 import styled from '@emotion/styled';
 import { Grid } from '@mui/material';
-import { FC, useEffect } from 'react';
+import { FC, useEffect, useRef } from 'react';
 import { Controller, useFieldArray, useForm } from 'react-hook-form';
 
+import { useAutoAnimate } from '../../../hooks/useAutoAnimate';
 import Completion from '../../../models/Completion';
 import CompletionQuantityTextField from '../completions/CompletionQuantityTextField';
 import CompletionUnitTextField from '../completions/CompletionUnitTextField';
@@ -36,6 +37,8 @@ const ValidateCompletionsForm: FC<Props> = ({ completions, formId, onSubmit }) =
     control, // control props comes from useForm
     name: 'completions'
   });
+  const animateRef = useRef(null);
+  useAutoAnimate(animateRef);
 
   useEffect(() => {
     const previousCompletions = getValues().completions;
@@ -68,7 +71,7 @@ const ValidateCompletionsForm: FC<Props> = ({ completions, formId, onSubmit }) =
   };
 
   return (
-    <form id={formId} onSubmit={handleSubmit(formatAndSubmit)}>
+    <form id={formId} onSubmit={handleSubmit(formatAndSubmit)} ref={animateRef}>
       {fields.map((field, index) => (
         <FieldsetGrid
           columns={2}